[IMP] Move spreadsheet data fields into spreadsheet_abstract and add init logic to set default locale of new sheets
