+++ spreadsheet_oca/models/spreadsheet_abstract.py
@@ -1,9 +1,10 @@
 # Copyright 2022 CreuBlanca
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
+import base64
 import json
 
-from odoo import api, fields, models
+from odoo import _, api, fields, models
 from odoo.exceptions import AccessError
 
 
@@ -12,12 +13,64 @@ class SpreadsheetAbstract(models.AbstractModel):
     _description = "Spreadsheet abstract for inheritance"
 
     name = fields.Char(required=True)
-    spreadsheet_raw = fields.Serialized()
+    spreadsheet_binary_data = fields.Binary(
+        string="Spreadsheet file",
+        default=lambda self: self._empty_spreadsheet_data_base64(),
+    )
+    spreadsheet_raw = fields.Serialized(
+        inverse="_inverse_spreadsheet_raw", compute="_compute_spreadsheet_raw"
+    )
     spreadsheet_revision_ids = fields.One2many(
         "spreadsheet.oca.revision",
         inverse_name="res_id",
         domain=lambda r: [("model", "=", r._name)],
     )
+
+    @api.depends("spreadsheet_binary_data")
+    def _compute_spreadsheet_raw(self):
+        for dashboard in self:
+            if dashboard.spreadsheet_binary_data:
+                dashboard.spreadsheet_raw = json.loads(
+                    base64.decodebytes(dashboard.spreadsheet_binary_data).decode(
+                        "UTF-8"
+                    )
+                )
+            else:
+                dashboard.spreadsheet_raw = {}
+
+    def _inverse_spreadsheet_raw(self):
+        for record in self:
+            record.spreadsheet_binary_data = base64.encodebytes(
+                json.dumps(record.spreadsheet_raw).encode("UTF-8")
+            )
+
+    def _empty_spreadsheet_data_base64(self):
+        """Create an empty spreadsheet workbook.
+        Encoded as base64
+        """
+        data = json.dumps(self._empty_spreadsheet_data())
+        return base64.b64encode(data.encode())
+
+    def _empty_spreadsheet_data(self):
+        """Create an empty spreadsheet workbook.
+        The sheet name should be the same for all users to allow consistent references
+        in formulas. It is translated for the user creating the spreadsheet.
+        """
+        lang = self.env["res.lang"]._lang_get(self.env.user.lang)
+        locale = lang._odoo_lang_to_spreadsheet_locale()
+        return {
+            "version": 1,
+            "sheets": [
+                {
+                    "id": "sheet1",
+                    "name": _("Sheet1"),
+                }
+            ],
+            "settings": {
+                "locale": locale,
+            },
+            "revisionId": "START_REVISION",
+        }
 
     def get_spreadsheet_data(self):
         self.ensure_one()
